优化审核通过更新业务

+++ src/main/java/local/tux/app/service/oa/TakeVacationService.java
@@ -87,7 +87,7 @@ public class TakeVacationService {
 		TakeVacation t = takeVacationRepository.findOneByUserId(user.get().getId());
 		if (t == null)
 			return 0;
-		return t.getUsable();
+		return t.getUsable().intValue();
 	}
 
 }
